test(webfont-generator): split incremental tests
